LayoutTests/css3/unicode-bidi-plaintext-parse.html

<!DOCTYPE HTML>
<link rel="stylesheet" href="../fast/js/resources/js-test-style.css">
<script src="../fast/js/resources/js-test-pre.js"></script>
<p id="description"></p><div id="console"></div>
<script>
description("Basic test for unicode-bidi: plaintext parsing.")
var testDiv = document.createElement("div");
testDiv.style.cssText = "unicode-bidi: -webkit-plaintext";
shouldBeEqualToString("testDiv.style.unicodeBidi", "-webkit-plaintext");
successfullyParsed = true;
</script>
<script src="../fast/js/resources/js-test-post.js"></script>
