LayoutTests/css3/unicode-bidi-plaintext-parse-expected.txt

Basic test for unicode-bidi: plaintext parsing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDiv.style.unicodeBidi is "-webkit-plaintext"
PASS successfullyParsed is true

TEST COMPLETE
